Update manage user permission

+++ moon/actions/auth.py
@@ -1,6 +1,6 @@
 from rest_framework.response import Response
 
-from rest_framework.authentication import SessionAuthentication, BasicAuthentication
+from rest_framework.authentication import SessionAuthentication
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 
@@ -9,20 +9,37 @@ from django.contrib.auth import authenticate
 from rest_framework.authtoken.models import Token
 from rest_framework.status import (
     HTTP_400_BAD_REQUEST,
+    HTTP_403_FORBIDDEN,
     HTTP_404_NOT_FOUND,
     HTTP_200_OK
 )
+
+
+class AdminUser(APIView):
+    authentication_classes = (SessionAuthentication, TokenAuthentication)
+    permission_classes = (IsAuthenticated,)
+
+    def get(self, request, format=None):
+        if request.user is None:
+            return None
+
+        if not request.user.is_superuser:
+            return Response(status=HTTP_403_FORBIDDEN)
+
+        data = {
+            'username': request.user.username,
+        }
+
+        return Response(data, status=HTTP_200_OK)
+
 
 class User(APIView):
     authentication_classes = (SessionAuthentication, TokenAuthentication)
     # permission_classes = (IsAuthenticated,)
 
     def get(self, request, format=None):
-        if request.user is None:
-            return None
-
         data = {
-            'username': request.user.username,
+            'username': request.user.username if request.user else 'Hi, Guest',
         }
 
         return Response(data, status=HTTP_200_OK)
@@ -54,6 +71,8 @@ class Logout(APIView):
     authentication_classes = (SessionAuthentication, TokenAuthentication)
 
     def post(self, request, format=None):
-        request.user.auth_token.delete()
+        if request.user and hasattr(request.user, 'auth_token'):
+            request.user.auth_token.delete()
+
         return Response(status=HTTP_200_OK)
